+++ scalyr_agent/builtin_monitors/syslog_monitor.py
@@ -17,12 +17,13 @@
 from __future__ import unicode_literals
 from __future__ import absolute_import
 
+from __future__ import print_function
 from scalyr_agent import compat
+from scalyr_agent.builtin_monitors.syslog_utils import SyslogLogConfigManager
 
 __author__ = "imron@scalyr.com"
 
 import errno
-import glob
 import logging
 import logging.handlers
 import os
@@ -37,7 +38,6 @@ from string import Template
 from io import open
 
 import six
-from six.moves import range
 import six.moves.socketserver
 
 from scalyr_agent import (
@@ -46,9 +46,6 @@ from scalyr_agent import (
     AutoFlushingRotatingFileHandler,
 )
 from scalyr_agent.monitor_utils.server_processors import RequestSizeExceeded
-from scalyr_agent.monitor_utils.auto_flushing_rotating_file import (
-    AutoFlushingRotatingFile,
-)
 from scalyr_agent.util import (
     StoppableThread,
     get_parser_from_config,
@@ -59,10 +56,7 @@ from scalyr_agent.scalyr_monitor import BadMonitorConfiguration
 docker_module_available = True
 
 try:
-    from scalyr_agent.builtin_monitors.docker_monitor import (
-        get_attributes_and_config_from_labels,
-        DockerOptions,
-    )
+    from scalyr_agent.builtin_monitors.docker_monitor import DockerOptions
 except ImportError:
     # Should typically not happen when using the docker mode because the Docker images we publish have this module
     # installed
@@ -110,7 +104,7 @@ define_config_option(
     "are stored. The file will be placed in the default Scalyr log directory, unless it is an "
     "absolute path",
     convert_to=six.text_type,
-    default="agent_syslog.log",
+    default="agent_syslog${HASH}.log",
 )
 
 define_config_option(
@@ -246,7 +240,7 @@ define_config_option(
     "the path is not absolute, then it is assumed to be relative to the main Scalyr Agent log "
     "directory.",
     convert_to=six.text_type,
-    default="containers/${CNAME}.log",
+    default="containers/${CNAME}${HASH}.log",
 )
 
 define_config_option(
@@ -680,106 +674,6 @@ class SyslogTCPServer(
         return False
 
 
-class LogDeleter(object):
-    """Deletes unused log files that match a log_file_template"""
-
-    def __init__(
-        self,
-        check_interval_mins,
-        delete_interval_hours,
-        check_rotated_timestamps,
-        max_log_rotations,
-        log_path,
-        log_file_template,
-    ):
-        self._check_interval = check_interval_mins * 60
-        self._delete_interval = delete_interval_hours * 60 * 60
-        self._check_rotated_timestamps = check_rotated_timestamps
-        self._max_log_rotations = max_log_rotations
-        self._log_glob = os.path.join(
-            log_path, log_file_template.safe_substitute(CID="*", CNAME="*")
-        )
-
-        self._last_check = time.time()
-
-    def _get_old_logs_for_glob(
-        self, current_time, glob_pattern, existing_logs, check_rotated, max_rotations
-    ):
-
-        result = []
-
-        for matching_file in glob.glob(glob_pattern):
-            try:
-                added = False
-                mtime = os.path.getmtime(matching_file)
-                if (
-                    current_time - mtime > self._delete_interval
-                    and matching_file not in existing_logs
-                ):
-                    result.append(matching_file)
-                    added = True
-
-                for i in range(max_rotations, 0, -1):
-                    rotated_file = matching_file + (".%d" % i)
-                    try:
-                        if not os.path.isfile(rotated_file):
-                            continue
-
-                        if check_rotated:
-                            mtime = os.path.getmtime(rotated_file)
-                            if current_time - mtime > self._delete_interval:
-                                result.append(rotated_file)
-                        else:
-                            if added:
-                                result.append(rotated_file)
-
-                    except OSError as e:
-                        global_log.warn(
-                            "Unable to read modification time for file '%s', %s"
-                            % (rotated_file, six.text_type(e)),
-                            limit_once_per_x_secs=300,
-                            limit_key="mtime-%s" % rotated_file,
-                        )
-
-            except OSError as e:
-                global_log.warn(
-                    "Unable to read modification time for file '%s', %s"
-                    % (matching_file, six.text_type(e)),
-                    limit_once_per_x_secs=300,
-                    limit_key="mtime-%s" % matching_file,
-                )
-        return result
-
-    def check_for_old_logs(self, existing_logs):
-
-        old_logs = []
-        current_time = time.time()
-        if current_time - self._last_check > self._check_interval:
-
-            old_logs = self._get_old_logs_for_glob(
-                current_time,
-                self._log_glob,
-                existing_logs,
-                self._check_rotated_timestamps,
-                self._max_log_rotations,
-            )
-            self._last_check = current_time
-
-        for filename in old_logs:
-            try:
-                os.remove(filename)
-                global_log.log(
-                    scalyr_logging.DEBUG_LEVEL_1, "Deleted old log file '%s'" % filename
-                )
-            except OSError as e:
-                global_log.warn(
-                    "Error deleting old log file '%s', %s"
-                    % (filename, six.text_type(e)),
-                    limit_once_per_x_secs=300,
-                    limit_key="delete-%s" % filename,
-                )
-
-
 class SyslogHandler(object):
     """Protocol neutral class for handling messages that come in from a syslog server
 
@@ -792,12 +686,16 @@ class SyslogHandler(object):
         logger,
         line_reporter,
         config,
+        global_config,
+        log_manager,
         server_host,
         log_path,
         get_log_watcher,
         rotate_options,
         docker_options,
     ):
+        self._global_config = global_config
+        self._config = config
 
         docker_logging = config.get("mode") == "docker"
         self.__docker_regex = None
@@ -831,14 +729,6 @@ class SyslogHandler(object):
             self.__docker_file_template = Template(
                 config.get("docker_logfile_template")
             )
-            self.__docker_log_deleter = LogDeleter(
-                config.get("docker_check_for_unused_logs_mins"),
-                config.get("docker_delete_unused_logs_hours"),
-                config.get("docker_check_rotated_timestamps"),
-                rotation_count,
-                log_path,
-                self.__docker_file_template,
-            )
 
             if config.get("docker_use_daemon_to_resolve"):
                 from scalyr_agent.builtin_monitors.docker_monitor import (
@@ -871,6 +761,14 @@ class SyslogHandler(object):
         self.__max_log_rotations = rotation_count
         self.__max_log_size = max_log_size
         self.__flush_delay = config.get("log_flush_delay")
+
+        self.__log_watcher = None
+        self.__log_manager = log_manager
+
+    def __del__(self):
+        self.__log_manager.close()
+        for cname in self.__docker_loggers:
+            self.__docker_loggers[cname].close()
 
     def __get_regex(self, config, field_name):
         value = config.get(field_name)
@@ -922,26 +820,6 @@ class SyslogHandler(object):
             raise
 
         return attributes
-
-    def __create_log_file(self, cname, cid, log_config):
-        """create our own rotating logger which will log raw messages out to disk.
-        """
-        result = None
-        try:
-            result = AutoFlushingRotatingFile(
-                filename=log_config["path"],
-                max_bytes=self.__max_log_size,
-                backup_count=self.__max_log_rotations,
-                flush_delay=self.__flush_delay,
-            )
-
-        except Exception as e:
-            global_log.error(
-                "Unable to open SyslogMonitor log file: %s" % six.text_type(e)
-            )
-            result = None
-
-        return result
 
     def __extract_container_info(self, data):
         """Attempts to extract the container id, container name and container labels from the log line received from Docker via
@@ -1030,7 +908,6 @@ class SyslogHandler(object):
     def __handle_docker_logs(self, data):
 
         watcher = None
-        module = None
         # log watcher for adding/removing logs from the agent
         if self.__get_log_watcher:
             watcher, module = self.__get_log_watcher()
@@ -1040,82 +917,48 @@ class SyslogHandler(object):
         if cname is None:
             return
 
-        current_time = time.time()
-        current_log_files = []
         self.__logger_lock.acquire()
         try:
-            logger = None
             if cname is not None and cid is not None:
                 # check if we already have a logger for this container
                 # and if not, then create it
                 if cname not in self.__docker_loggers:
-                    info = dict()
-
-                    attrs, base_config = get_attributes_and_config_from_labels(
-                        labels, self._docker_options
+                    # the log manager uses the "message_log" config to name files, so we take the docker file template,
+                    # sub in known values, and pass that along as "message_log"
+                    modified_config = {}
+                    modified_config.update(self._config)
+                    modified_config.update(
+                        {
+                            "message_log": self.__docker_file_template.safe_substitute(
+                                {"CNAME": cname, "CID": cid}
+                            ),
+                            "parser": "agentSyslogDocker",
+                            "containerName": cname,
+                            "containerId": cid,
+                        }
                     )
+                    if self.__server_host:
+                        modified_config["serverHost"] = self.__server_host
 
-                    # get the config and set the attributes
-                    info["log_config"] = self.__create_log_config(
-                        cname, cid, base_config, attrs
+                    self.__docker_loggers[cname] = SyslogLogConfigManager(
+                        self._global_config,
+                        None,
+                        self.__max_log_size,
+                        self.__max_log_rotations,
+                        extra_config=modified_config,
                     )
-                    info["cid"] = cid
+                    self.__docker_loggers[cname].set_log_watcher(watcher)
+            app, host = self._get_app_and_host(data)
+            logger = self.__docker_loggers[cname].get_logger(app)
+            print(cname)
+            print(app)
 
-                    # create the physical log files
-                    info["logger"] = self.__create_log_file(
-                        cname, cid, info["log_config"]
-                    )
-                    info["last_seen"] = current_time
-
-                    # if we created the log file
-                    if info["logger"]:
-                        # add it to the main scalyr log watcher
-                        if watcher and module:
-                            info["log_config"] = watcher.add_log_config(
-                                module.module_name, info["log_config"]
-                            )
-
-                        # and keep a record for ourselves
-                        self.__docker_loggers[cname] = info
-                    else:
-                        global_log.warn("Unable to create logger for %s." % cname)
-                        return
-
-                # at this point __docker_loggers will always contain
-                # a logger for this container name, so log the message
-                # and mark the time
-                logger = self.__docker_loggers[cname]
-                logger["last_seen"] = current_time
-
-            if self.__expire_count >= RUN_EXPIRE_COUNT:
-                self.__expire_count = 0
-
-                # find out which if any of the loggers in __docker_loggers have
-                # expired
-                expired = []
-
-                for key, info in six.iteritems(self.__docker_loggers):
-                    if current_time - info["last_seen"] > self.__docker_expire_log:
-                        expired.append(key)
-
-                # remove all the expired loggers
-                for key in expired:
-                    info = self.__docker_loggers.pop(key, None)
-                    if info:
-                        info["logger"].close()
-                        if watcher and module:
-                            watcher.remove_log_path(
-                                module.module_name, info["log_config"]["path"]
-                            )
-            self.__expire_count += 1
-
-            for key, info in six.iteritems(self.__docker_loggers):
-                current_log_files.append(info["log_config"]["path"])
+            # TODO: recreate `if self.__expire_count >= RUN_EXPIRE_COUNT:` section
         finally:
             self.__logger_lock.release()
 
         if logger:
-            logger["logger"].write(line_content)
+            logger.info(data)
         else:
             global_log.warning(
                 "Syslog writing docker logs to syslog file instead of container log",
@@ -1124,8 +967,24 @@ class SyslogHandler(object):
             )
             self.__logger.info(data)
 
-        if self.__docker_log_deleter:
-            self.__docker_log_deleter.check_for_old_logs(current_log_files)
+        for cname in self.__docker_loggers:
+            self.__docker_loggers[cname].check_for_old_logs()
+
+    @staticmethod
+    def _get_app_and_host(data):
+        app = "unknown"
+        host = "unknown"
+
+        splitdata = data.split(" ")
+        if len(splitdata) > 3:
+            if splitdata[3].endswith(":"):
+                app = splitdata[3].split("[")[0]
+            else:
+                host = splitdata[3]
+                if len(splitdata) > 4:
+                    app = splitdata[4].split("[")[0]
+
+        return app, host
 
     def handle(self, data):  # type: (six.text_type) -> None
         """
@@ -1138,9 +997,14 @@ class SyslogHandler(object):
         if self.__docker_logging:
             self.__handle_docker_logs(data)
         else:
-            self.__logger.info(data)
+            app, host = self._get_app_and_host(data)
+            logger = self.__log_manager.get_logger(app)
+            logger.info(data)
         # We add plus one because the calling code strips off the trailing new lines.
         self.__line_reporter(data.count("\n") + 1)
+
+    def set_log_watcher(self, log_watcher):
+        self.__log_manager.set_log_watcher(log_watcher)
 
 
 class RequestVerifier(object):
@@ -1184,6 +1048,8 @@ class SyslogServer(object):
         port,
         logger,
         config,
+        global_config,
+        log_manager,
         line_reporter,
         accept_remote=False,
         server_host=None,
@@ -1248,6 +1114,8 @@ class SyslogServer(object):
             logger,
             line_reporter,
             config,
+            global_config,
+            log_manager,
             server_host,
             log_path,
             get_log_watcher,
@@ -1279,6 +1147,9 @@ class SyslogServer(object):
             # need to think of what to do for 2.4, which will hang on shutdown when run as standalone
             if hasattr(server, "shutdown"):
                 run_state.register_on_stop_callback(server.shutdown)
+
+    def set_log_watcher(self, watcher):
+        self.__server.syslog_handler.set_log_watcher(watcher)
 
     def start(self, run_state):
         self.__prepare_run_state(run_state)
@@ -1379,6 +1250,26 @@ running. You can find this log file in the [Overview](/logStart) page. By defaul
                 "This may be due to not including the docker library when building container image.",
                 "mode",
             )
+        if (
+            not self._config.get("mode") == "docker"
+            and "${HASH}" not in self._config.get("message_log")
+            and len(self._global_config.syslog_log_configs) > 0
+        ):
+            raise BadMonitorConfiguration(
+                "Failing `syslog_logs` is defined in the but `message_log` does not contain a template placeholder "
+                "`${HASH}`. This is required for syslog_logs configurations to function correctly.",
+                "message_log",
+            )
+        if (
+            self._config.get("mode") == "docker"
+            and "${HASH}" not in self._config.get("docker_logfile_template")
+            and len(self._global_config.syslog_log_configs) > 0
+        ):
+            raise BadMonitorConfiguration(
+                "Failing `syslog_logs` is defined in the but `docker_logfile_template` does not contain a template placeholder "
+                "`${HASH}`. This is required for syslog_logs configurations to function correctly.",
+                "docker_logfile_template",
+            )
 
         # the main server
         self.__server = None
@@ -1413,17 +1304,7 @@ running. You can find this log file in the [Overview](/logStart) page. By defaul
         # configure the logger and path
         self.__message_log = self._config.get("message_log")
 
-        self.log_config = {
-            "parser": self._config.get("parser"),
-            "path": self.__message_log,
-        }
-
         self.__flush_delay = self._config.get("log_flush_delay")
-        try:
-            attributes = JsonObject({"monitor": "agentSyslog"})
-            self.log_config["attributes"] = attributes
-        except Exception:
-            global_log.error("Error setting monitor attribute in SyslogMonitor")
 
         (
             default_rotation_count,
@@ -1439,6 +1320,16 @@ running. You can find this log file in the [Overview](/logStart) page. By defaul
             self.__max_log_rotations = default_rotation_count
 
         self._docker_options = None
+
+        self._log_manager = SyslogLogConfigManager(
+            self._global_config,
+            None,
+            self.__max_log_size,
+            self.__max_log_rotations,
+            extra_config=self._config,
+        )
+
+        self.log_config = self._log_manager.get_config(".*")
 
     def __build_server_list(self, protocol_string):
         """Builds a list containing (protocol, port) tuples, based on a comma separated list
@@ -1585,6 +1476,8 @@ running. You can find this log file in the [Overview](/logStart) page. By defaul
                 protocol[1],
                 self.__disk_logger,
                 self._config,
+                self._global_config,
+                self._log_manager,
                 line_reporter,
                 accept_remote=self.__accept_remote_connections,
                 server_host=self.__server_host,
@@ -1593,6 +1486,7 @@ running. You can find this log file in the [Overview](/logStart) page. By defaul
                 rotate_options=rotate_options,
                 docker_options=self._docker_options,
             )
+            self.__server.set_log_watcher(self.__log_watcher)
 
             # iterate over the remaining items creating servers for each protocol
             for p in self.__server_list[1:]:
@@ -1601,6 +1495,8 @@ running. You can find this log file in the [Overview](/logStart) page. By defaul
                     p[1],
                     self.__disk_logger,
                     self._config,
+                    self._global_config,
+                    self._log_manager,
                     line_reporter,
                     accept_remote=self.__accept_remote_connections,
                     server_host=self.__server_host,
@@ -1609,6 +1505,7 @@ running. You can find this log file in the [Overview](/logStart) page. By defaul
                     rotate_options=rotate_options,
                     docker_options=self._docker_options,
                 )
+                server.set_log_watcher(self.__log_watcher)
                 self.__extra_servers.append(server)
 
             # start any extra servers in their own threads